requirements.txt
PyQt6
python-chess
pyyaml
fastapi
uvicorn
pydantic
